Rename detail-route-utils to loader-utils
