+++ src/routes/-detail-route-loader-utils.ts
